# web/openapi.json
{"components":{"schemas":{"CurrentUserBody":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["https://example.com/api/schemas/CurrentUserBody.json"],"format":"uri","readOnly":true,"type":"string"},"auth_enabled":{"description":"Whether single sign-on is enabled on this instance.","type":"boolean"},"authenticated":{"type":"boolean"},"email":{"examples":["user@example.com"],"type":"string"},"username":{"examples":["user"],"type":"string"}},"required":["email","username","authenticated","auth_enabled"],"type":"object"},"ErrorDetail":{"additionalProperties":false,"properties":{"location":{"description":"Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'","type":"string"},"message":{"description":"Error message text","type":"string"},"value":{"description":"The value at the given location"}},"type":"object"},"ErrorModel":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["https://example.com/api/schemas/ErrorModel.json"],"format":"uri","readOnly":true,"type":"string"},"detail":{"description":"A human-readable explanation specific to this occurrence of the problem.","examples":["Property foo is required but is missing."],"type":"string"},"errors":{"description":"Optional list of individual error details","items":{"$ref":"#/components/schemas/ErrorDetail"},"type":["array","null"]},"instance":{"description":"A URI reference that identifies the specific occurrence of the problem.","examples":["https://example.com/error-log/abc123"],"format":"uri","type":"string"},"status":{"description":"HTTP status code","examples":[400],"format":"int64","type":"integer"},"title":{"description":"A short, human-readable summary of the problem type. This value should not change between occurrences of the error.","examples":["Bad Request"],"type":"string"},"type":{"default":"about:blank","description":"A URI reference to human-readable documentation for the error.","examples":["https://example.com/errors/example"],"format":"uri","type":"string"}},"type":"object"},"JobView":{"additionalProperties":false,"properties":{"cron":{"type":"string"},"disable_cron":{"type":"boolean"},"disable_fail_fast":{"type":"boolean"},"disabled":{"type":"boolean"},"name":{"type":"string"},"retries":{"format":"int64","type":"integer"},"runs":{"items":{"$ref":"#/components/schemas/RunView"},"type":["array","null"]},"slug":{"type":"string"},"timeout":{"format":"int64","type":"integer"}},"required":["runs","name","slug","cron","disable_cron","disable_fail_fast","disabled"],"type":"object"},"ListLogsByRunIDsRow":{"additionalProperties":false,"properties":{"created_at":{"format":"int64","type":"integer"},"created_at_time":{},"message":{"type":"string"},"run_id":{"format":"int64","type":"integer"},"severity_id":{"format":"int64","type":"integer"}},"required":["created_at","run_id","severity_id","message","created_at_time"],"type":"object"},"LogoutResponseBody":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["https://example.com/api/schemas/LogoutResponseBody.json"],"format":"uri","readOnly":true,"type":"string"},"logout_url":{"description":"Where to send the browser to clear the provider-side session. Empty when the provider does not advertise an end_session_endpoint.","type":"string"}},"type":"object"},"RunView":{"additionalProperties":false,"properties":{"duration":{"type":"string"},"end_time":{"type":"string"},"id":{"format":"int64","type":"integer"},"job_name":{"type":"string"},"logs":{"items":{"$ref":"#/components/schemas/ListLogsByRunIDsRow"},"type":["array","null"]},"start_time":{"type":"string"},"start_time_unix":{"format":"int64","type":"integer"},"status_id":{"format":"int64","type":"integer"}},"required":["id","job_name","status_id","start_time_unix","start_time","end_time","duration","logs"],"type":"object"}},"securitySchemes":{"cookieAuth":{"description":"Session cookie issued by /api/auth/callback.","in":"cookie","name":"gocron_session","type":"apiKey"}}},"info":{"title":"GoCron API","version":"dev"},"openapi":"3.1.0","paths":{"/api/auth/logout":{"post":{"description":"Revokes the current session and clears the session cookie.","operationId":"post-auth-logout","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutResponseBody"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"security":[{"cookieAuth":[]}],"summary":"Logout","tags":["Auth"]}},"/api/auth/me":{"get":{"description":"Returns whether single sign-on is enabled and the identity of the current session.","operationId":"get-auth-me","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CurrentUserBody"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"security":[{"cookieAuth":[]}],"summary":"Get current user","tags":["Auth"]}},"/api/command":{"post":{"description":"Execute command.","operationId":"post-command","requestBody":{"content":{"multipart/form-data":{"encoding":{"command":{"contentType":"text/plain"}},"schema":{"properties":{"command":{"description":"command to execute","maxLength":255,"minLength":1,"type":"string"}},"required":["command"],"type":"object"}}}},"responses":{"204":{"description":"No Content"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"security":[{"cookieAuth":[]}],"summary":"Execute command","tags":["Command"]}},"/api/jobs":{"get":{"description":"Get jobs with run details but no logs.","operationId":"get-jobs","responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/JobView"},"type":["array","null"]}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"security":[{"cookieAuth":[]}],"summary":"Get jobs","tags":["Jobs"]},"post":{"description":"Start all jobs in order of name.","operationId":"post-jobs","responses":{"204":{"description":"No Content"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"security":[{"cookieAuth":[]}],"summary":"Start jobs","tags":["Jobs"]},"put":{"description":"Change single job.","operationId":"put-job","parameters":[{"description":"job name","explode":false,"in":"query","name":"name","schema":{"description":"job name","maxLength":255,"type":"string"}},{"description":"action to perform","explode":false,"in":"query","name":"action","schema":{"description":"action to perform","enum":["toggle","disable_all","enable_all","enable_scheduled","enable_non_scheduled"],"type":"string"}}],"responses":{"204":{"description":"No Content"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"security":[{"cookieAuth":[]}],"summary":"Change job","tags":["Jobs"]}},"/api/jobs/{name}":{"post":{"description":"Start single job.","operationId":"post-job","parameters":[{"description":"job name","in":"path","name":"name","required":true,"schema":{"description":"job name","maxLength":255,"type":"string"}}],"responses":{"204":{"description":"No Content"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"security":[{"cookieAuth":[]}],"summary":"Start job","tags":["Jobs"]}},"/api/runs/{job_name}":{"get":{"description":"Get runs with logs for a job.","operationId":"get-runs","parameters":[{"description":"job name","in":"path","name":"job_name","required":true,"schema":{"description":"job name","maxLength":255,"minLength":1,"type":"string"}},{"description":"number of runs to return","explode":false,"in":"query","name":"limit","schema":{"default":5,"description":"number of runs to return","format":"int64","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/RunView"},"type":["array","null"]}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"security":[{"cookieAuth":[]}],"summary":"Get runs","tags":["Runs"]}}},"servers":[{"url":""}]}